배너 등록 테스트 코드 작성

+++ src/main/java/green/study/domain/lecture/model/Lecture.java
@@ -10,7 +10,7 @@ public class Lecture {
 
     private Long key;
     private String title;
-    private Double price;
+    private int price;
     private String description;
     private LectureImage image;
     private Long memberKey;
